+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 * Bugfix: Need css style name is now based on need-type and not on the longer, whitespace-containing type name.
   Example: ``need-test`` instead of not valid ``need-test case``. (`#108 <https://github.com/useblocks/sphinxcontrib-needs/issues/108>`_)
 * Bugfix: No more exception raise if ``copy`` value not set inside :ref:`needs_extra_links`.
+* Improvement: Better log message, if required id is missing. (`#112 <https://github.com/useblocks/sphinxcontrib-needs/issues/112>`_)
 
 0.4.3
 -----